% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLog.R
\name{aggregateStats}
\alias{aggregateStats}
\title{Do statistics on multiple logs}
\usage{
aggregateStats(lmatch, statFun = mean)
}
\arguments{
\item{lmatch}{a list of logs produced by \code{\link{getLog}}}

\item{statFun}{The statistical function that you want to use on the logs.
  Default is a simple mean, but can include any function, R or user made,
  with a \code{na.rm} parameter}
}
\value{
A matrix of the model appliled log table for all the participants
  involved in any of the games. There is an extra column of \code{gp}, or
  games played.
}
\description{
Do statistics on multiple logs
}

