% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tf2statr.R
\docType{data}
\name{eventArchive}
\alias{eventArchive}
\title{Archive of tf2 events and their corresponding log ID}
\format{A list of the different events in variable "events" \describe{
  \item{nameOfFirstLevel}{The comp.tf webpage name last portion}
  \item{nameOfSecondLevel}{The different games in the event with designations
  in the naming scheme. For tournaments, it is sequenced as the following:the
  name of the round as given by comp.tf, the match number from the top of the
  bracket and then the map number in that match.} \item{secondLevel}{The log
  ID is a string of the log.tf log ID} }}
\source{
\url{http://comp.tf/wiki/Main_Page} with the function that scrapes it
  and saves new entires the archive \code{\link{getLogIDsComptf}}
}
\description{
A dictionary for the comp.tf website so you don't have to re scrape it
everytime you want to examine a data set. The raw data is in JSON format
while the '.R' code translates it into a list of lists using 'jsonlite'
package.
}

