% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLogIDs.R
\name{getLogIDsComptf}
\alias{getLogIDsComptf}
\title{Scrap comp.tf event pages for logs IDs}
\usage{
getLogIDsComptf(comptfToken, scrapeLoc = c("Tourney", "Season")[1],
  withNames = TRUE)
}
\arguments{
\item{comptfToken}{The name of the comp.tf page in the url}

\item{scrapeLoc}{Is this page a tourney (scrape a bracket) or a season
(scrape a table)}

\item{withNames}{Should the identifier of each of the logs be added as a
  names for output?}
}
\value{
A character vector with all the logs. A named character vector if
  "withNames" is TRUE
}
\description{
Scrap comp.tf event pages for logs IDs
}

