% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLogIDs.R
\name{getLogIDsJSON}
\alias{getLogIDsJSON}
\title{Use logs.tf's query JSON API}
\usage{
getLogIDsJSON(title = "", uploader = "", player = "", num = 10)
}
\arguments{
\item{title}{The title of the match that you want to query}

\item{uploader}{The person or bot that actually put the game on log.tf}

\item{player}{The SteamID3 [U:int:numbers] or the 16 digit id}

\item{num}{The maximum number of IDs to output}
}
\value{
a vector of IDs
}
\description{
The log.tf query api has three queries that can be made, an uploader, a name
and a player. Out of these, the player is the most reliable without prior knowledge.
The result is a vector of log IDs that can be then feed into 'getLog'
One of the queries are required but all three can be supplied
}

