% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tf2statr.R
\docType{data}
\name{playerDict}
\alias{playerDict}
\title{Archive of player names and their corresponding Steam ID3}
\format{A data frame, "playerDict" with 2 columns and an ever expanding
  number of rows:
\describe{
  \item{name}{default is the custom profile vanity
  name as a string}
  \item{steamID3}{The Steam ID3 of the player in the form of
  [u:int:numbers] as a character} }}
\source{
the redirect from the Steam ID3 from Valve. The function that creates
  more rows using this method is \code{convSID2name}
}
\description{
A dictionary for the program to provide alternative names for the logs
instead of the steam ID3 which is unknown. Names have to be relatively
constant so results can be compared across games and timeperiods.
}

