% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/queryLogstf.R
\name{queryLogstf}
\alias{queryLogstf}
\title{Do complicated queries of TF2stats}
\usage{
queryLogstf(players = c(), teamName = "", season = "", tournament = "",
  shGetLog = TRUE)
}
\arguments{
\item{players}{A vector of players either given as Steam ID3s or custom
profile names}

\item{teamName}{TODO}

\item{season}{The string of the comp.tf webpage end piece (Insomnia52)}

\item{tournament}{The string of the comp.tf webpage end piece (Insomnia52)}

\item{shGetLog}{Should the program go ahead and get the logs of the query or
 just output the log IDs.}
}
\value{
Either a vector of log IDs if shGetLog is false, or a list of the
 relevant logs
}
\description{
Builds a list of logs that are relevant for a query using the union of all the
log IDs for the individual queries. The queries that can be done right now are
player queries using log.tf Search tool as well as season and tournament using
comp.tf historical pages. The players can be inputed as a list of Steam ID3s
or custom profile names which are then converted into steam ID3s. Whenever
possible, the program uses the archive to find log IDs for events.
}

