% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tf2statr.R
\docType{package}
\name{tf2star}
\alias{tf2star}
\alias{tf2star-package}
\title{tf2statr: Managing and querying Team Fortress 2 stats}
\description{
This package provides tools to query logs.tf for stats and attempts to provide
advanced queries and cleaning functions using key TF2 sites.
}

