% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLog.R
\name{getLog}
\alias{getLog}
\title{Grabs the raw JSON of the log and clean it up}
\usage{
getLog(logId, useAltNames = FALSE, keepClassSpec = FALSE,
  keepChat = FALSE)
}
\arguments{
\item{logId}{The sequential ID that logs.tf uses for a given game}

\item{useAltNames}{should the program use the alternative names from the
player dictionary in 'data/playerDict.csv' or go find the alt name using
their steamID and their custom profile name?}

\item{keepClassSpec}{Do you want to keep the class specific stats for
everyone?}

\item{keepChat}{Do you want to keep the chat from the game?}
}
\value{
A list with the logs, the player specific stats in $player and the
	 table version of the numeric stats in $table with some added stats
}
\description{
Grabs the raw JSON of the log and clean it up
}

